requirements.txt
streamlit
streamlit-option-menu
watchdog
openpyxl
sqlalchemy
pyodbc
psycopg2-binary
streamlit-aggrid